Modified signature method

+++ src/main/java/com/sdk/accumulate/model/AddCreditsArg.java
@@ -1,11 +1,13 @@
 package com.sdk.accumulate.model;
 
 
+import java.math.BigInteger;
+
 public class AddCreditsArg {
 
     private String recipient;
 
-    private int amount;
+    private BigInteger amount;
 
     public String getRecipient() {
         return recipient;
@@ -15,11 +17,11 @@ public class AddCreditsArg {
         this.recipient = recipient;
     }
 
-    public int getAmount() {
+    public BigInteger getAmount() {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(BigInteger amount) {
         this.amount = amount;
     }
 }
