+++ src/main/java/com/sdk/accumulate/model/BurnTokensArg.java
@@ -1,14 +1,16 @@
 package com.sdk.accumulate.model;
 
+import java.math.BigInteger;
+
 public class BurnTokensArg {
 
-    private int amount;
+    private BigInteger amount;
 
-    public int getAmount() {
+    public BigInteger getAmount() {
         return amount;
     }
 
-    public void setAmount(int amount) {
+    public void setAmount(BigInteger amount) {
         this.amount = amount;
     }
 }
